move tensor stats & decode cache to re_viewer_context
