+++ crates/re_viewer_context/Cargo.toml
@@ -19,6 +19,7 @@ all-features = true
 re_arrow_store.workspace = true
 re_data_store = { workspace = true, features = ["serde"] }
 re_log_types = { workspace = true, features = ["ecolor", "glam", "image"] }
+re_log.workspace = true
 re_query.workspace = true
 re_renderer.workspace = true
 re_ui.workspace = true
@@ -28,12 +29,14 @@ anyhow.workspace = true
 bytemuck.workspace = true
 egui.workspace = true
 glam.workspace = true
+half.workspace = true
 itertools.workspace = true
+lazy_static.workspace = true
+ndarray.workspace = true
 nohash-hasher.workspace = true
 serde = "1"
 slotmap.workspace = true
 uuid = { version = "1.1", features = ["serde", "v4", "js"] }
-lazy_static.workspace = true
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 puffin.workspace = true
