+++ crates/re_viewer_context/src/tensor/mod.rs
@@ -1,9 +1,9 @@
-mod mesh_cache;
+// TODO(andreas): Move tensor utilities to a tensor specific crate?
+
 mod tensor_decode_cache;
 mod tensor_stats;
 mod tensor_stats_cache;
 
-pub use mesh_cache::MeshCache;
 pub use tensor_decode_cache::TensorDecodeCache;
 pub use tensor_stats::TensorStats;
 pub use tensor_stats_cache::TensorStatsCache;
