+++ crates/re_viewer_context/src/tensor/tensor_stats.rs
